Cleans up snackbar context provider
