+++ frontend/src/App.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react'
+import React from 'react'
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom'
 import theme from './theme'
 import { ThemeProvider } from '@mui/material/styles'
@@ -9,38 +9,20 @@ import CreateElectionTemplates from './components/ElectionForm/CreateElectionTem
 import Election from './components/Election/Election'
 import Sandbox from './components/Sandbox'
 import LandingPage from './components/LandingPage'
-import { Alert, Box, CssBaseline, Snackbar } from '@mui/material'
-import { Isnack, SnackbarContext } from './components/SnackbarContext'
+import { Box, CssBaseline } from '@mui/material'
+import { SnackbarContextProvider } from './components/SnackbarContext'
 import Footer from './components/Footer'
 import { ConfirmDialogProvider } from './components/ConfirmationDialogProvider'
 import About from './components/About'
 import { AuthSessionContextProvider } from './components/AuthSessionContextProvider'
 
 const App = () => {
-  // const authSession = useAuthSession()
-  const [snack, setSnack] = useState({
-    message: '',
-    severity: "info",
-    open: false,
-    autoHideDuration: null,
-  } as Isnack)
-  const handleClose = (event?: React.SyntheticEvent | Event, reason?: string) => {
-    if (reason === 'clickaway') {
-      return;
-    }
-    setSnack({ ...snack, open: false })
-  }
   return (
     <Router>
       <ThemeProvider theme={theme}>
         <AuthSessionContextProvider>
           <ConfirmDialogProvider>
-            <SnackbarContext.Provider value={{ snack, setSnack }}>
-              <Snackbar open={snack.open} autoHideDuration={snack.autoHideDuration} onClose={handleClose} anchorOrigin={{ vertical: 'bottom', horizontal: 'center' }}>
-                <Alert severity={snack.severity} onClose={handleClose}>
-                  {snack.message}
-                </Alert>
-              </Snackbar>
+            <SnackbarContextProvider>
               <CssBaseline />
               <Box display='flex' flexDirection='column' minHeight={'100vh'} >
                 <Header />
@@ -60,7 +42,7 @@ const App = () => {
                 </Box>
                 <Footer />
               </Box>
-            </SnackbarContext.Provider>
+            </SnackbarContextProvider>
           </ConfirmDialogProvider>
         </AuthSessionContextProvider>
       </ThemeProvider>
