+++ frontend/src/components/SnackbarContext.tsx
@@ -1,5 +1,6 @@
-import React from 'react'
+import React, { useContext, useState } from 'react'
 import { createContext, Dispatch, SetStateAction } from 'react'
+import { Alert, Snackbar } from '@mui/material'
 
 export interface Isnack {
     message: string,
@@ -13,14 +14,46 @@ export interface ISnackBarContext {
     setSnack: Dispatch<SetStateAction<Isnack>>
 }
 
-
 export const SnackbarContext = createContext<ISnackBarContext>({
     snack: {
         message: '',
         severity: "info",
         open: false,
         autoHideDuration: null,
-    } as Isnack,
+    },
     setSnack: () => false
 }
 )
+
+export const SnackbarContextProvider = ({ children }) => {
+
+    const [snack, setSnack] = useState<Isnack>({
+        message: '',
+        severity: "info",
+        open: false,
+        autoHideDuration: null,
+    })
+
+    const handleClose = (event?: React.SyntheticEvent | Event, reason?: string) => {
+        if (reason === 'clickaway') {
+            return;
+        }
+        setSnack({ ...snack, open: false })
+    }
+
+    return (
+        <SnackbarContext.Provider value={{ snack, setSnack }}>
+            <Snackbar open={snack.open} autoHideDuration={snack.autoHideDuration} onClose={handleClose} anchorOrigin={{ vertical: 'bottom', horizontal: 'center' }}>
+                <Alert severity={snack.severity} onClose={handleClose}>
+                    {snack.message}
+                </Alert>
+            </Snackbar>
+            {children}
+        </SnackbarContext.Provider>
+    )
+}
+
+
+export default function useSnackbar() {
+    return useContext(SnackbarContext);
+}
